Always reset environment + database data in system test case, not just integration test case.
